refactor(RestaurantDetailView): 컴포넌트 디렉터리 계층 수정 및 css 분리
