+++ src/App.tsx
@@ -3,7 +3,7 @@ import Header from './components/Header';
 import FilterContainer from './components/FilterContainter';
 import RestaurantList from './components/RestaurantList';
 import Modal from './components/Modal';
-import RestaurantDetailView from './components/Modal/RestaurantDetailView';
+import RestaurantDetailView from './components/RestaurantDetailView';
 
 import { Restaurant } from './types';
 import mockData from './mockData.json';
